Use own components for Area and Mountains layers
