+++ src/components/yeti/map-layers/layer.js
@@ -1,3 +1,5 @@
+import ol from '@/js/libs/ol';
+
 export default {
   computed: {
     map() {
@@ -6,6 +8,15 @@ export default {
     },
     view() {
       return this.map.getView();
+    },
+  },
+  methods: {
+    getExtent(projection) {
+      let extent = this.view.calculateExtent(this.map.getSize() || null);
+      if (projection) {
+        extent = ol.proj.transformExtent(extent, ol.proj.get('EPSG:3857'), ol.proj.get(projection));
+      }
+      return extent;
     },
   },
 };
